Page changes

+++ backend/controllers/PageController.php
@@ -4,23 +4,28 @@ namespace backend\controllers;
 
 use Yii;
 use common\models\Page;
-use backend\models\search\PageSearch;
-use yii\web\Controller;
+use common\models\search\PageSearch;
+use common\components\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use yii\helpers\Json;
+use yii\helpers\Url;
 
 /**
  * PageController implements the CRUD actions for Page model.
  */
 class PageController extends Controller
 {
+    /**
+     * @inheritdoc
+     */
     public function behaviors()
     {
         return [
             'verbs' => [
                 'class' => VerbFilter::className(),
                 'actions' => [
-                    'delete' => ['post'],
+                    'delete' => ['POST'],
                 ],
             ],
         ];
@@ -32,6 +37,25 @@ class PageController extends Controller
      */
     public function actionIndex()
     {
+        if (Yii::$app->request->post('hasEditable')) {
+            if (!Yii::$app->user->can('/' . \common\components\helpers\Url::normalizeRoute('update'))) {
+                throw new HttpException(403);
+            }
+
+            $model = $this->findModel(Yii::$app->request->post('editableKey'));
+
+            $out = Json::encode(['output'=>'', 'message'=>'']);
+
+            $posted = current($_POST['Page']);
+            $post = ['Page' => $posted];
+
+            if ($model->load($post)) {
+                $model->save();
+            }
+            echo $out;
+            return;
+        }
+
         $searchModel = new PageSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -39,6 +63,26 @@ class PageController extends Controller
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
         ]);
+    }
+
+    /**
+     * Displays a single Page model.
+     * @param integer $id
+     * @return mixed
+     */
+    public function actionView($id)
+    {
+        $model=$this->findModel($id);
+
+        if ($model->load(Yii::$app->request->post()) && $model->save()) {
+            if (!Yii::$app->user->can('/' . \common\components\helpers\Url::normalizeRoute('update'))) {
+                throw new HttpException(403);
+            }
+            Yii::$app->session->setFlash('kv-detail-success', 'Saved record successfully');
+            return $this->redirect(['view', 'id' => $model->id]);
+        } else {
+            return $this->render('view', ['model'=>$model]);
+        }
     }
 
     /**
@@ -51,7 +95,7 @@ class PageController extends Controller
         $model = new Page();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['index']);
+            return $this->redirect(['index', 'returned'=>true]);
         } else {
             return $this->render('create', [
                 'model' => $model,
@@ -70,7 +114,7 @@ class PageController extends Controller
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['index']);
+            return $this->redirect(['index', 'returned'=>true]);
         } else {
             return $this->render('update', [
                 'model' => $model,
@@ -84,11 +128,31 @@ class PageController extends Controller
      * @param integer $id
      * @return mixed
      */
-    public function actionDelete($id)
-    {
-        $this->findModel($id)->delete();
-
-        return $this->redirect(['index']);
+    public function actionDelete($id) {
+        $post = Yii::$app->request->post();
+        if (Yii::$app->request->isAjax && isset($post['j-delete'])) {
+            if ($this->findModel($id)->delete()) {
+                echo Json::encode([
+                    'success' => true,
+                    'messages' => [
+                    'kv-detail-info' => 'Successfully deleted. <a href="' .
+                                                Url::to(['index']) . '" class="btn btn-sm btn-info">' .
+                        '<i class="glyphicon glyphicon-hand-right"></i>  Click here</a> to proceed.'
+                    ]
+                ]);
+            } else {
+                echo Json::encode([
+                    'success' => false,
+                    'messages' => [
+                        'kv-detail-error' => 'Cannot delete'
+                    ]
+                ]);
+            }
+            return;
+        } else {
+            $this->findModel($id)->delete();
+            $this->redirect(['index', 'returned'=>true]);
+        }
     }
 
     /**
